+++ include/visionaray/math/detail/vector4.inl
@@ -121,14 +121,14 @@ inline T const* vector<4, T>::data() const
 
 template <typename T>
 MATH_FUNC
-inline T& vector<4, T>::operator[](size_t i)
+inline T& vector<4, T>::operator[](unsigned i)
 {
     return data()[i];
 }
 
 template <typename T>
 MATH_FUNC
-inline T const& vector<4, T>::operator[](size_t i) const
+inline T const& vector<4, T>::operator[](unsigned i) const
 {
     return data()[i];
 }
